+++ src/python/pants/backend/shell/shell_command_test.py
@@ -22,9 +22,8 @@ from pants.backend.shell.target_types import (
 from pants.core.goals.run import RunRequest
 from pants.core.target_types import ArchiveTarget, FilesGeneratorTarget, FileSourceField
 from pants.core.target_types import rules as core_target_type_rules
-from pants.core.util_rules.archive import rules as archive_rules
+from pants.core.util_rules import source_files, system_binaries
 from pants.core.util_rules.source_files import SourceFiles, SourceFilesRequest
-from pants.core.util_rules.source_files import rules as source_files_rules
 from pants.engine.addresses import Address
 from pants.engine.fs import EMPTY_SNAPSHOT, DigestContents
 from pants.engine.process import Process
@@ -41,9 +40,9 @@ from pants.testutil.rule_runner import QueryRule, RuleRunner
 def rule_runner() -> RuleRunner:
     rule_runner = RuleRunner(
         rules=[
-            *archive_rules(),
+            *system_binaries.rules(),
             *shell_command_rules(),
-            *source_files_rules(),
+            *source_files.rules(),
             *core_target_type_rules(),
             QueryRule(GeneratedSources, [GenerateFilesFromShellCommandRequest]),
             QueryRule(Process, [ShellCommandProcessRequest]),
